+++ src/main/resources/application.properties
@@ -447,6 +447,7 @@ surname-normalized.filiol = filol
 surname-normalized.fiscina = fiscine, fisini, fisina
 surname-normalized.fondovila = fondevila
 surname-normalized.forastiere = forastieri, forestiere, forestieri
+surname-normalized.frescaroli = frascaroli
   # matches: Forastiere, Forastieri, Forestiere, Forestieri
 surname-normalized.gailour = gailours, gailur, gailurs, galour, galours, galur, galurs
 surname-normalized.gamio = degamio
@@ -463,10 +464,12 @@ surname-normalized.hesaine = hesaines
 surname-normalized.hipolito = ipolito
 surname-normalized.hiriart = hiriarte, iriart, iriarte
 surname-normalized.iacovone = iacovoni, giacovone, giacovoni
+surname-normalized.iparaguire = deiparaguire
 surname-normalized.iratcaval = iratsaval, iracaval, irasaval
   # matches: Iratcabal, Iratssabal, Iracabal, Irassabal
 surname-normalized.irivuru = hirivuru, hirivourou
 surname-normalized.ituralde = ithuralde, ithouralde
+surname-normalized.ituria = deituria
 surname-normalized.iturios = ituriots
 surname-normalized.ituriosgarai = ituriotsgarai
 surname-normalized.iturvide = ithurvide
@@ -495,6 +498,7 @@ surname-normalized.marturano = martorano
 surname-normalized.mcave = macave
   # matches: Mac Cabe, MacCabe, Mc Cabe, McCabe, Mac Cave, Macabe
 surname-normalized.medio = demedio
+surname-normalized.milione = miglione, melione, meglione
 surname-normalized.mociaro = mocero
   # matches: Mocciaro, Moccero
 surname-normalized.montagna = montana
