+++ src/test/java/com/geneaazul/gedcomanalyzer/service/familytree/PlainFamilyTreeServiceTests.java
@@ -1,10 +1,11 @@
-package com.geneaazul.gedcomanalyzer.service;
+package com.geneaazul.gedcomanalyzer.service.familytree;
 
 import com.geneaazul.gedcomanalyzer.config.GedcomAnalyzerProperties;
+import com.geneaazul.gedcomanalyzer.mapper.RelationshipMapper;
 import com.geneaazul.gedcomanalyzer.model.EnrichedPerson;
-import com.geneaazul.gedcomanalyzer.model.Relationship;
+import com.geneaazul.gedcomanalyzer.model.FormattedRelationship;
 import com.geneaazul.gedcomanalyzer.model.Relationships;
-import com.geneaazul.gedcomanalyzer.service.familytree.NetworkFamilyTreeService;
+import com.geneaazul.gedcomanalyzer.service.PersonService;
 import com.geneaazul.gedcomanalyzer.service.storage.GedcomHolder;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,54 +23,41 @@ import java.util.List;
 @SpringBootTest
 @EnableConfigurationProperties
 @ActiveProfiles("test")
-public class NetworkFamilyTreeServiceTests {
+@SuppressWarnings("DataFlowIssue")
+public class PlainFamilyTreeServiceTests {
 
     @Autowired
     private GedcomHolder gedcomHolder;
     @Autowired
     private PersonService personService;
     @Autowired
-    private NetworkFamilyTreeService networkFamilyTreeService;
+    private PlainFamilyTreeService plainFamilyTreeService;
+    @Autowired
+    private RelationshipMapper relationshipMapper;
     @Autowired
     private GedcomAnalyzerProperties properties;
 
     @Test
-    public void testGenerateNetworkHTML() throws IOException {
+    public void testExportToPDF() throws IOException {
 
         EnrichedPerson person = gedcomHolder.getGedcom().getPersonById("I4");
         List<Relationships> relationshipsList = personService.setTransientProperties(person, false);
 
         MutableInt index = new MutableInt(1);
-        List<EnrichedPerson> peopleInTree = relationshipsList
+        List<FormattedRelationship> peopleInTree = relationshipsList
                 .stream()
                 .map(Relationships::findLast)
                 .sorted()
-                .limit(100)
-                .map(Relationship::person)
-                .peek(p -> p.setOrderKey(index.getAndIncrement()))
+                .peek(relationship -> relationship.person().setOrderKey(index.getAndIncrement()))
+                .map(relationship -> relationshipMapper.toRelationshipDto(relationship, false))
+                .map(relationship -> relationshipMapper.formatInSpanish(relationship, true))
                 .toList();
 
-        Path htmlPyvisNetworkFilePath = properties
+        Path path = properties
                 .getTempDir()
                 .resolve("family-trees")
-                .resolve("export_to_html_test.html");
-
-        Path csvPyvisNetworkNodesFilePath = properties
-                .getTempDir()
-                .resolve("family-trees")
-                .resolve("export_to_csv_nodes_test.csv");
-
-        Path csvPyvisNetworkEdgesFilePath = properties
-                .getTempDir()
-                .resolve("family-trees")
-                .resolve("export_to_csv_edges_test.csv");
-
-        networkFamilyTreeService.generateNetworkHTML(
-                htmlPyvisNetworkFilePath,
-                csvPyvisNetworkNodesFilePath,
-                csvPyvisNetworkEdgesFilePath,
-                false,
-                peopleInTree);
+                .resolve("export_to_pdf_test.pdf");
+        plainFamilyTreeService.exportToPDF(path, person, peopleInTree);
     }
 
 }
